+++ src/main/java/de/fernunihagen/dna/scalephant/performance/osm/OSMStructureCallback.java
@@ -1,0 +1,13 @@
+package de.fernunihagen.dna.scalephant.performance.osm;
+
+import de.fernunihagen.dna.scalephant.performance.osm.util.GeometricalStructure;
+
+public interface OSMStructureCallback {
+
+	/**
+	 * Process the given geometrical structure
+	 * @param geometricalStructure
+	 */
+	public void processStructure(final GeometricalStructure geometricalStructure);
+	
+}