+++ changelog.md
@@ -5,6 +5,7 @@
 - Improvement: Implemented keep alive packages, to keep tcp connections open
 - Improvement: Added the possibility to log debug messages 
 - Improvement: Write geometrical data to server in OSMInsertBenchmark
+- Improvement: Moved OSM importer code to own class
 - Bugfix: Removed duplicate logging entries
 - Bugfix: Removed deadlock on node down event
 - Bugfix: Skip compact operation when only one SSTable exists